delete Contribute Module. Fixes in ConsultDoc modules
