--- UDV-Core/examples/DemoContribute.html
@@ -1,18 +1,0 @@
-<html>
-<head>
-    <title>UDV-Core</title>
-    <link rel="stylesheet" href="../src/css/styles.css">
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-</head>
-<body>
-    <div id="viewerDiv"></div>
-    <img id="logoIMU" src="../img/logo-imu.png"></img>
-    <img id="logoLIRIS" src="../img/logo-liris.png"></img>
-    <script src="../dist/udvcore.js"></script>
-    <script src="../node_modules/moment/moment.js"></script>
-    <script src="../node_modules/three/build/three.js"></script>
-    <script src="../src/Externals/dat.gui/dat.gui.min.js"></script>
-    <script src="DemoContribute.js"></script>
-</body>
-</html>
