+++ UDV-Core/src/Main.js
@@ -14,5 +14,3 @@ export { MiniMapController }    from './Modules/Others/MiniMap';
 export { TemporalController }   from './Modules/Temporal/Temporal';
 
 export { DocumentController}    from './Modules/ConsultDoc/DocumentController';
-
-export { ContributeController}    from './Extensions/Contribute/src/ContributeController';
